test: rename test suites and tests to drop `test_asdf_` prefix
It was just annoying, and mostly unnecessary as long as none of the
names clash with existing symbols.  Dropping these prefixes makes the
the logs a bit more readable for me and also makes it less hassle to
type individual test names.

Did however keep `test_asdf_` in the case of unit tests for specific
functions like `test_asdf_mapping_get`, etc.