chore: first pass at implementing a YAML Pointer parser
The YAML Pointer path is represented as an STC vector of
asdf_yaml_path_component_t which I think is useful.

Considered writing an implementation that doesn't parse the full path at
once, but just iterates over it, but the problem with that is that it
wouldn't necessarily detect from the start if the path is invalid, so we
parse the full path in one pass.

gh-104